Create Tournament_Winners.java
There are N knights numbered [1, 2, .., N]. A tournament is organized among these knights. This tournament will have N-1 matches and at every match, the winner stays in and the looser moves out. There are no draws in the match. At the end of the tournament, there is only one knight left- the winner. The spymaster has some information about who would win in a one-to-one match between two players, i.e., for some pairs of knights, he know who would win in a one-to-one match between them. Now he wants to know if a scheduling of matches can help someone to win the tournament. That is for some knight x, whether it is possible to play only the matches where he will win. He wants you to find out that for how many knights, such kind of schedule is possible.