Remove nodes from Path sum less than B
Given a binary tree with root node pointer A and an integer B. A complete path is defined as a path from the root to a leaf. The sum of all nodes on that path is defined as the sum of that path. You have to remove (prune the tree) all nodes which don't lie in any path with sum >= B.